tests: Extend test_roundtrip
Check the case when, via `dbWriteTable`, we are writing a data.frame with columns
ordered differently than the table we are writing to.  In `dbWriteTable` we
attempt to describe parameters (types, length) of the INSERT query using
information about the columns of the table being written to.  Some care is needed
to make sure that the table column descriptions get paired with the appropriate parameters.